extractTable
