+++ content-scripts/modules/utilities/extractTable.ts
